# openapi/agent-api.yaml
openapi: 3.1.0
info:
  title: Busibox Agent Server API
  version: 1.0.0
  description: |
    Production-grade agent server with Pydantic AI for executing AI agents with tool calls,
    dynamic agent/workflow management, and OAuth2 token forwarding to Busibox services.
    
    ## Authentication
    All endpoints (except `/health`) require Bearer token authentication with Busibox JWT.
    
    JWT tokens are issued by the AuthZ service (`authz-lxc:8010`) using OAuth2 token exchange.
    This service validates JWTs using the AuthZ JWKS endpoint (`/.well-known/jwks.json`).
    See `openapi/authz-api.yaml` for token issuance details.
    
    ## Features
    - Dynamic agent definition management with personal/built-in agents
    - Tool and workflow CRUD operations
    - Agent run execution with SSE streaming
    - Intelligent query routing via dispatcher
    - Performance evaluation with scorers
    - Schedule management for cron-based runs
    
  contact:
    name: Busibox Infrastructure Team
    email: infra@busibox.local
servers:
  - url: http://agent-lxc:8000
    description: Production (agent-lxc container)
  - url: http://10.96.200.30:8000
    description: Test environment
  - url: http://localhost:8000
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /dispatcher/route:
    post:
      summary: Route user query
      description: |
        Route a natural language query to appropriate tools and agents.
        
        Analyzes the query and returns:
        - Selected tools/agents for routing
        - Confidence score (0-1)
        - Reasoning for the decision
        - Alternative suggestions
        
        The routing decision is logged for accuracy measurement and debugging.
      tags: [dispatcher]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatcherRequest'
      responses:
        '200':
          description: Routing decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatcherResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /insights/init:
    post:
      summary: Initialize insights collection
      description: |
        Initialize the chat_insights collection in Milvus.
        Creates the collection with schema if it doesn't exist.
        This endpoint is idempotent - safe to call multiple times.
      tags: [insights]
      responses:
        '200':
          description: Collection initialized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  collection:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /insights:
    post:
      summary: Insert insights
      description: Insert chat insights into Milvus for RAG
      tags: [insights]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsertInsightsRequest'
      responses:
        '200':
          description: Insights inserted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /insights/search:
    post:
      summary: Search insights
      description: Search for relevant insights based on query
      tags: [insights]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsightSearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsightSearchResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /insights/conversation/{conversation_id}:
    delete:
      summary: Delete conversation insights
      description: Delete all insights for a conversation
      tags: [insights]
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Insights deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  conversationId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /insights/user/{user_id}:
    delete:
      summary: Delete user insights
      description: Delete all insights for a user (for account deletion/cleanup)
      tags: [insights]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Insights deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  userId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /insights/stats/{user_id}:
    get:
      summary: Get user insight statistics
      description: Get statistics for user insights
      tags: [insights]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsightStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /insights/flush:
    post:
      summary: Flush insights collection
      description: Flush collection to ensure data persistence
      tags: [insights]
      responses:
        '200':
          description: Collection flushed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  collection:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Busibox JWT token
...
  schemas:
...
    DispatcherRequest:
      type: object
      properties:
        query:
          type: string
          description: Natural language query
          example: Search for documents about machine learning
        available_tools:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
        available_agents:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              description:
                type: string
        attachments:
          type: array
          items:
            type: object
          default: []
      required:
        - query
        - available_tools
        - available_agents
    DispatcherResponse:
      type: object
      properties:
        selected_tools:
          type: array
          items:
            type: string
          description: Names of selected tools
        selected_agents:
          type: array
          items:
            type: string
            format: uuid
          description: IDs of selected agents
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Confidence score for routing decision
        reasoning:
          type: string
          description: Explanation for the routing decision
        alternatives:
          type: array
          items:
            type: object
            properties:
              tools:
                type: array
                items:
                  type: string
              agents:
                type: array
                items:
                  type: string
                  format: uuid
              confidence:
                type: number
                format: float
      required:
        - selected_tools
        - selected_agents
        - confidence
        - reasoning
...
    ChatInsight:
      type: object
      properties:
        id:
          type: string
          description: Insight ID
        userId:
          type: string
          description: User ID who owns this insight
        content:
          type: string
          description: The insight text
          maxLength: 5000
        embedding:
          type: array
          items:
            type: number
            format: float
          description: Vector embedding (1024 dimensions)
        conversationId:
          type: string
          description: Source conversation ID
        analyzedAt:
          type: integer
          description: Unix timestamp when insight was extracted
      required:
        - id
        - userId
        - content
        - embedding
        - conversationId
        - analyzedAt
    InsertInsightsRequest:
      type: object
      properties:
        insights:
          type: array
          items:
            $ref: '#/components/schemas/ChatInsight'
          minItems: 1
          description: List of insights to insert
      required:
        - insights
    InsightSearchRequest:
      type: object
      properties:
        query:
          type: string
          description: Search query
          minLength: 1
          maxLength: 500
        userId:
          type: string
          description: User ID to filter results
        limit:
          type: integer
          description: Maximum number of results
          minimum: 1
          maximum: 20
          default: 3
        scoreThreshold:
          type: number
          format: float
          description: Maximum L2 distance threshold
          minimum: 0.0
          maximum: 2.0
          default: 0.7
      required:
        - query
        - userId
    InsightSearchResult:
      type: object
      properties:
        id:
          type: string
          description: Insight ID
        userId:
          type: string
          description: User ID
        content:
          type: string
          description: Insight content
        conversationId:
          type: string
          description: Conversation ID
        analyzedAt:
          type: string
          format: date-time
          description: ISO timestamp
        score:
          type: number
          format: float
          description: Similarity score (L2 distance)
      required:
        - id
        - userId
        - content
        - conversationId
        - analyzedAt
        - score
    InsightSearchResponse:
      type: object
      properties:
        query:
          type: string
          description: Original query
        results:
          type: array
          items:
            $ref: '#/components/schemas/InsightSearchResult'
          description: Search results
        count:
          type: integer
          description: Number of results returned
      required:
        - query
        - results
        - count
    InsightStatsResponse:
      type: object
      properties:
        userId:
          type: string
          description: User ID
        count:
          type: integer
          description: Number of insights
        collectionName:
          type: string
          description: Collection name
      required:
        - userId
        - count
        - collectionName
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
      required:
        - detail
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'